Move mapreduce logic to DataLayouts layer
